requirements.txt
streamlit
groq
chromadb
sentence-transformers
pymupdf
python-dotenv